Move standard json definition to its own file
These structures are defined in different places.
